初次提交

+++ PickerDemo/PinPad/PPCircleButton.m
@@ -10,8 +10,7 @@
 
 @implementation PPCircleButton
 
-- (id)initWithFrame:(CGRect)frame
-{
+- (id)initWithFrame:(CGRect)frame {
     self = [super initWithFrame:frame];
     if (self) {
         // Initialization code
@@ -23,23 +22,21 @@
 // Only override drawRect: if you perform custom drawing.
 // An empty implementation adversely affects performance during animation.
 
-- (void)drawRect:(CGRect)rect
-{
+- (void)drawRect:(CGRect)rect {
     // Drawing code
     [super drawRect:rect];
-    [self.layer setCornerRadius:CGRectGetHeight(rect)/2.0];
+    [self.layer setCornerRadius:(CGFloat) (CGRectGetHeight(rect) / 2.0)];
     self.layer.borderColor = [[UIColor whiteColor] colorWithAlphaComponent:0.5].CGColor;
     self.layer.borderWidth = 2.0f;
 }
 
 - (void)setHighlighted:(BOOL)highlighted {
-    
-    if(highlighted) {
+
+    if (highlighted) {
         self.layer.borderColor = [self titleColorForState:UIControlStateHighlighted].CGColor;
-    }
-    else {
+    } else {
         self.layer.borderColor = [[UIColor whiteColor] colorWithAlphaComponent:0.5].CGColor;
     }
-    
+
 }
 @end
